Initial layout for the saved pipeline parameters
