+++ src/main/webapp/resources/js/pages/launch/components/ParametersModal/index.js
@@ -1,0 +1,3 @@
+import { ParametersModal } from "./ParametersModal";
+
+export default ParametersModal;
